// web/src/components/Skills/logoAvailable.type.ts
type logoAvailable =
    "Kotlin"
    | "Python"
    | "PHP"
    | "React Native"
    | "ReactJS"
    | "CodeIgniter"
    | "CSS"
    | "HTML"
    | "JavaScript"
    | "NodeJS"
    | "TypeScript"
    | "Git"
    | "MongoDB"
    | "MySQL"
    | "Oracle Database"
    | "OCaml"
    | "TailwindCSS"
    | "Android Studio"
    | "Api RESTful"
    | "Golang"
    | "SQL";


export default logoAvailable;